+++ CityWatch.Web/Pages/roster/GuardRosterAction.cshtml
@@ -1,0 +1,5 @@
+@page
+@model CityWatch.Web.Pages.roster.GuardRosterActionModel
+@{
+    Layout = null;
+}
