+++ CityWatch.Web/Program.cs
@@ -68,6 +68,7 @@ builder.Services.AddScoped<ISmsGlobalService, SmsGlobalService>();
 builder.Services.AddScoped<ISignalRNotificationService, SignalRNotificationService>();
 builder.Services.AddScoped<ITimesheetReportGenerator, TimesheetReportGenerator>();
 builder.Services.AddScoped<IRosterReportGenerator, RosterReportGenerator>();
+builder.Services.AddScoped<IGuardRosterReportGenerator, GuardRosterReportGenerator>();
 builder.Services.AddScoped<ILogbookDataService, LogbookDataService>();
 builder.Services.AddScoped<ICertificateGenerator, CertificateGenerator>();
 builder.Services.AddScoped<IRPLCertificateGeneratorService, RPLCertificateGeneratorService>();
